+++ mbo/container/limited_map.h
@@ -25,15 +25,11 @@
 #include <type_traits>
 #include <utility>
 
-#include "absl/log/absl_log.h"
 #include "mbo/container/internal/limited_ordered.h"
 #include "mbo/types/compare.h"
 #include "mbo/types/traits.h"
 
 namespace mbo::container {
-
-// NOLINTNEXTLINE(cppcoreguidelines-macro-usage)
-#define LV_REQUIRE(severity, condition) ABSL_LOG_IF(severity, !(condition))
 
 // NOLINTBEGIN(readability-identifier-naming)
 
@@ -385,8 +381,6 @@ constexpr auto ToLimitedMap(std::pair<K, V> (&&array)[N], const KComp& key_comp 
 // NOLINTEND(*-avoid-c-arrays)
 
 // NOLINTEND(readability-identifier-naming)
-
-#undef LV_REQUIRE
 
 }  // namespace mbo::container